Update nav.rst
I don't know if it was always this way, but as of the current version of Flask-Nav, you actually need to call the decorator with the ID of your nav element, as you want it to be registered on the `nav` context global. A tiny thing really, but this minor detail tripped me up at first.