Add more block flexibility to base.html
Forgive me if I've overlooked a simpler way to do this, but I was looking for a good way to inject my own wrapper divs into my base.html (inheriting bootstrap/base.html) and ended up making these changes to allow it. This PR would allow a user's base template t do things like this:

```
{% block before_content %}
{# render pagination #}
{% endblock before_content %}

{% block outer_content %}
<div class="container container-content">
  {{super()}}
</div>
{% endblock %}

{% block outer_body %}
<div class="container container-body">
  {{super()}}
</div>
{% endblock %}

{% block scripts %}
<script>
// Now I can use the intuitive "scripts" block without nuking the core scripts.
</script>
{% endblock scripts %}
```

I realize that some of this flexibility can be achieved by careful use of super(), but I prefer not to feel so forced to use it in my page templates. In any case, this is a very useful little extension, and I thank you for it! :)