Update forms.py
Wrapped inputs get rendered without name and so the form object does not work correctly. I'm not sure what is the correct fix for this, this looks to me more as a quickfix as probably more node's attributes should be put there. Or, why not just use the node render itself?
Even the sample app's example form does not work because of this.